+++ src/csi.ts
@@ -66,7 +66,7 @@ export default class CSI extends IndexFile {
     const nameSectionLength = bytes.readInt32LE(offset + 24)
 
     const { refIdToName, refNameToId } = this._parseNameBytes(
-      bytes.subarray(offset + 28, offset + 28 + nameSectionLength),
+      bytes.slice(offset + 28, offset + 28 + nameSectionLength),
     )
 
     return {
@@ -187,7 +187,7 @@ export default class CSI extends IndexFile {
   parsePseudoBin(bytes: Buffer, offset: number) {
     const lineCount = longToNumber(
       Long.fromBytesLE(
-        bytes.subarray(offset + 28, offset + 36) as unknown as number[],
+        bytes.slice(offset + 28, offset + 36) as unknown as number[],
         true,
       ),
     )