+++ src/index.ts
@@ -2,4 +2,5 @@ import TabixIndexedFile from './tabixIndexedFile'
 import TBI from './tbi'
 import CSI from './csi'
 
+console.log('ret')
 export { TabixIndexedFile, TBI, CSI }